Renaming and fixing warnings
